+++ src/lib/errors.ts
@@ -1,0 +1,7 @@
+export class AbortError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = 'AbortError'
+        this.message = message + ''
+    }
+}