+++ src/lib/braze/filters/index.ts
@@ -1,3 +1,4 @@
 import hash from './hash';
+import str from './string';
 
-export default { ...hash }
+export default { ...hash, ...str }