+++ src/lib/braze/filters/string.ts
@@ -1,0 +1,6 @@
+export default {
+    // usage: {{ 'abcde' | slice: 1, -1 }}
+    slice: (str: string, start: number, end: number) => {
+        return str.slice(start, end);
+    },
+}