+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [0.0.1] - 2019-04-10
 - Initial release
-- Preview simple Braze Liquid templates
-- Implemented filter: property_accessor
-- Implemented tag: connected_content
+- Implemented filter: 
+  - property_accessor
+- Implemented tag: 
+  - connected_content
+  - abort_message