+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.0.2] - 2019-04-11
+- Removed previewText
+- Re-implemented preview using Webview API
+- Added filter:
+  - slice
+
 ## [0.0.1] - 2019-04-10
 - Initial release
 - Implemented filter: